refactor: remove local DAL and interface
This turned out to be quite a lot of work for not much benefit at the
moment, it was a bit premature. Will reassess later.

Fixes #96
