+++ controlplane/internal/dal/dal.go
@@ -27,21 +27,149 @@ import (
 	"github.com/TBD54566975/ftl/internal/log"
 	"github.com/TBD54566975/ftl/internal/maps"
 	"github.com/TBD54566975/ftl/internal/slices"
+	ftlv1 "github.com/TBD54566975/ftl/protos/xyz/block/ftl/v1"
 	pschema "github.com/TBD54566975/ftl/protos/xyz/block/ftl/v1/schema"
 	"github.com/TBD54566975/ftl/schema"
 )
 
-var _ DAL = (*Postgres)(nil)
+var (
+	// ErrConflict is returned by select methods in the DAL when a resource already exists.
+	//
+	// Its use will be documented in the corresponding methods.
+	ErrConflict = errors.New("conflict")
+	// ErrNotFound is returned by select methods in the DAL when no results are found.
+	ErrNotFound = errors.New("not found")
+)
 
-func NewPostgres(conn sql.DBI) *Postgres {
-	return &Postgres{db: sql.NewDB(conn)}
+type DeploymentArtefact struct {
+	Digest     sha256.SHA256
+	Executable bool
+	Path       string
 }
 
-type Postgres struct {
+func (d *DeploymentArtefact) ToProto() *ftlv1.DeploymentArtefact {
+	return &ftlv1.DeploymentArtefact{
+		Digest:     d.Digest.String(),
+		Executable: d.Executable,
+		Path:       d.Path,
+	}
+}
+
+func DeploymentArtefactFromProto(in *ftlv1.DeploymentArtefact) (DeploymentArtefact, error) {
+	digest, err := sha256.ParseSHA256(in.Digest)
+	if err != nil {
+		return DeploymentArtefact{}, errors.WithStack(err)
+	}
+	return DeploymentArtefact{
+		Digest:     digest,
+		Executable: in.Executable,
+		Path:       in.Path,
+	}, nil
+}
+
+type Runner struct {
+	Key      model.RunnerKey
+	Language string
+	Endpoint string
+	State    RunnerState
+	// Assigned deployment key, if any.
+	Deployment types.Option[model.DeploymentKey]
+}
+
+type Reconciliation struct {
+	Deployment model.DeploymentKey
+	Module     string
+	Language   string
+
+	AssignedReplicas int
+	RequiredReplicas int
+}
+
+type RunnerState string
+
+// Runner states.
+const (
+	RunnerStateIdle     = RunnerState(sql.RunnerStateIdle)
+	RunnerStateReserved = RunnerState(sql.RunnerStateReserved)
+	RunnerStateAssigned = RunnerState(sql.RunnerStateAssigned)
+	RunnerStateDead     = RunnerState(sql.RunnerStateDead)
+)
+
+func RunnerStateFromProto(state ftlv1.RunnerState) RunnerState {
+	return RunnerState(strings.ToLower(state.String()))
+}
+
+func (s RunnerState) ToProto() ftlv1.RunnerState {
+	return ftlv1.RunnerState(ftlv1.RunnerState_value[strings.ToUpper(string(s))])
+}
+
+type DataPoint interface {
+	isDataPoint()
+}
+
+type MetricHistogram struct {
+	Count  int64
+	Sum    int64
+	Bucket []uint64
+}
+
+func (MetricHistogram) isDataPoint() {}
+
+type MetricCounter struct {
+	Value int64
+}
+
+func (MetricCounter) isDataPoint() {}
+
+type Metric struct {
+	RunnerKey  model.RunnerKey
+	StartTime  time.Time
+	EndTime    time.Time
+	SourceVerb schema.VerbRef
+	DestVerb   schema.VerbRef
+	Name       string
+	DataPoint  DataPoint
+}
+
+type Deployment struct {
+	Key         model.DeploymentKey
+	Language    string
+	Module      string
+	MinReplicas int
+	Schema      *schema.Module
+}
+
+type Status struct {
+	Runners     []Runner
+	Deployments []Deployment
+}
+
+// A Reservation of a Runner.
+type Reservation interface {
+	Runner() Runner
+	Commit(ctx context.Context) error
+	Rollback(ctx context.Context) error
+}
+
+func WithReservation(ctx context.Context, reservation Reservation, fn func() error) error {
+	if err := fn(); err != nil {
+		if rerr := reservation.Rollback(ctx); rerr != nil {
+			err = errors.Join(err, rerr)
+		}
+		return errors.WithStack(err)
+	}
+	return errors.WithStack(reservation.Commit(ctx))
+}
+
+func New(conn sql.DBI) *DAL {
+	return &DAL{db: sql.NewDB(conn)}
+}
+
+type DAL struct {
 	db *sql.DB
 }
 
-func (d *Postgres) GetStatus(ctx context.Context, allDeployments bool, allRunners bool) (Status, error) {
+func (d *DAL) GetStatus(ctx context.Context, allDeployments bool, allRunners bool) (Status, error) {
 	runners, err := d.db.GetActiveRunners(ctx, allRunners)
 	if err != nil {
 		return Status{}, errors.Wrap(translatePGError(err), "could not get active runners")
@@ -89,7 +217,7 @@ func (d *Postgres) GetStatus(ctx context.Context, allDeployments bool, allRunner
 	}, nil
 }
 
-func (d *Postgres) GetRunnersForDeployment(ctx context.Context, deployment model.DeploymentKey) ([]Runner, error) {
+func (d *DAL) GetRunnersForDeployment(ctx context.Context, deployment model.DeploymentKey) ([]Runner, error) {
 	runners := []Runner{}
 	rows, err := d.db.GetRunnersForDeployment(ctx, sqltypes.Key(deployment))
 	if err != nil {
@@ -107,13 +235,13 @@ func (d *Postgres) GetRunnersForDeployment(ctx context.Context, deployment model
 	return runners, nil
 }
 
-func (d *Postgres) UpsertModule(ctx context.Context, language, name string) (err error) {
+func (d *DAL) UpsertModule(ctx context.Context, language, name string) (err error) {
 	_, err = d.db.UpsertModule(ctx, language, name)
 	return errors.WithStack(translatePGError(err))
 }
 
 // GetMissingArtefacts returns the digests of the artefacts that are missing from the database.
-func (d *Postgres) GetMissingArtefacts(ctx context.Context, digests []sha256.SHA256) ([]sha256.SHA256, error) {
+func (d *DAL) GetMissingArtefacts(ctx context.Context, digests []sha256.SHA256) ([]sha256.SHA256, error) {
 	have, err := d.db.GetArtefactDigests(ctx, sha256esToBytes(digests))
 	if err != nil {
 		return nil, errors.WithStack(translatePGError(err))
@@ -125,7 +253,7 @@ func (d *Postgres) GetMissingArtefacts(ctx context.Context, digests []sha256.SHA
 }
 
 // CreateArtefact inserts a new artefact into the database and returns its ID.
-func (d *Postgres) CreateArtefact(ctx context.Context, content []byte) (digest sha256.SHA256, err error) {
+func (d *DAL) CreateArtefact(ctx context.Context, content []byte) (digest sha256.SHA256, err error) {
 	sha256digest := sha256.Sum(content)
 	_, err = d.db.CreateArtefact(ctx, sha256digest[:], content)
 	return sha256digest, errors.WithStack(translatePGError(err))
@@ -135,7 +263,7 @@ func (d *Postgres) CreateArtefact(ctx context.Context, content []byte) (digest s
 // previously created artefacts with it.
 //
 // If an existing deployment with identical artefacts exists, it is returned.
-func (d *Postgres) CreateDeployment(ctx context.Context, language string, schema *schema.Module, artefacts []DeploymentArtefact) (key model.DeploymentKey, err error) {
+func (d *DAL) CreateDeployment(ctx context.Context, language string, schema *schema.Module, artefacts []DeploymentArtefact) (key model.DeploymentKey, err error) {
 	// Start the transaction
 	tx, err := d.db.Begin(ctx)
 	if err != nil {
@@ -201,7 +329,7 @@ func (d *Postgres) CreateDeployment(ctx context.Context, language string, schema
 	return model.DeploymentKey(deploymentKey), nil
 }
 
-func (d *Postgres) GetDeployment(ctx context.Context, id model.DeploymentKey) (*model.Deployment, error) {
+func (d *DAL) GetDeployment(ctx context.Context, id model.DeploymentKey) (*model.Deployment, error) {
 	deployment, err := d.db.GetDeployment(ctx, sqltypes.Key(id))
 	if err != nil {
 		return nil, errors.WithStack(translatePGError(err))
@@ -213,7 +341,7 @@ func (d *Postgres) GetDeployment(ctx context.Context, id model.DeploymentKey) (*
 //
 // ErrConflict will be returned if a runner with the same endpoint and a
 // different key already exists.
-func (d *Postgres) UpsertRunner(ctx context.Context, runner Runner) error {
+func (d *DAL) UpsertRunner(ctx context.Context, runner Runner) error {
 	var pgDeploymentKey types.Option[sqltypes.Key]
 	if dkey, ok := runner.Deployment.Get(); ok {
 		pgDeploymentKey = types.Some(sqltypes.Key(dkey))
@@ -238,7 +366,7 @@ func (d *Postgres) UpsertRunner(ctx context.Context, runner Runner) error {
 }
 
 // KillStaleRunners deletes runners that have not had heartbeats for the given duration.
-func (d *Postgres) KillStaleRunners(ctx context.Context, age time.Duration) (int64, error) {
+func (d *DAL) KillStaleRunners(ctx context.Context, age time.Duration) (int64, error) {
 	count, err := d.db.KillStaleRunners(ctx, pgtype.Interval{
 		Microseconds: int64(age / time.Microsecond),
 		Valid:        true,
@@ -247,7 +375,7 @@ func (d *Postgres) KillStaleRunners(ctx context.Context, age time.Duration) (int
 }
 
 // DeregisterRunner deregisters the given runner.
-func (d *Postgres) DeregisterRunner(ctx context.Context, key model.RunnerKey) error {
+func (d *DAL) DeregisterRunner(ctx context.Context, key model.RunnerKey) error {
 	count, err := d.db.DeregisterRunner(ctx, sqltypes.Key(key))
 	if err != nil {
 		return errors.WithStack(translatePGError(err))
@@ -258,7 +386,7 @@ func (d *Postgres) DeregisterRunner(ctx context.Context, key model.RunnerKey) er
 	return nil
 }
 
-func (d *Postgres) ReserveRunnerForDeployment(ctx context.Context, language string, deployment model.DeploymentKey, reservationTimeout time.Duration) (Reservation, error) {
+func (d *DAL) ReserveRunnerForDeployment(ctx context.Context, language string, deployment model.DeploymentKey, reservationTimeout time.Duration) (Reservation, error) {
 	ctx, cancel := context.WithTimeout(ctx, reservationTimeout)
 	tx, err := d.db.Begin(ctx)
 	if err != nil {
@@ -310,7 +438,7 @@ func (p *postgresClaim) Rollback(ctx context.Context) error {
 func (p *postgresClaim) Runner() Runner { return p.runner }
 
 // SetDeploymentReplicas activates the given deployment.
-func (d *Postgres) SetDeploymentReplicas(ctx context.Context, key model.DeploymentKey, minReplicas int) error {
+func (d *DAL) SetDeploymentReplicas(ctx context.Context, key model.DeploymentKey, minReplicas int) error {
 	err := d.db.SetDeploymentDesiredReplicas(ctx, sqltypes.Key(key), int32(minReplicas))
 	if err != nil {
 		return errors.WithStack(translatePGError(err))
@@ -320,7 +448,7 @@ func (d *Postgres) SetDeploymentReplicas(ctx context.Context, key model.Deployme
 
 // GetDeploymentsNeedingReconciliation returns deployments that have a
 // mismatch between the number of assigned and required replicas.
-func (d *Postgres) GetDeploymentsNeedingReconciliation(ctx context.Context) ([]Reconciliation, error) {
+func (d *DAL) GetDeploymentsNeedingReconciliation(ctx context.Context) ([]Reconciliation, error) {
 	counts, err := d.db.GetDeploymentsNeedingReconciliation(ctx)
 	if err != nil {
 		if isNotFound(err) {
@@ -342,7 +470,7 @@ func (d *Postgres) GetDeploymentsNeedingReconciliation(ctx context.Context) ([]R
 // GetIdleRunnersForLanguage returns up to limit idle runners for the given language.
 //
 // If no runners are available, it will return an empty slice.
-func (d *Postgres) GetIdleRunnersForLanguage(ctx context.Context, language string, limit int) ([]Runner, error) {
+func (d *DAL) GetIdleRunnersForLanguage(ctx context.Context, language string, limit int) ([]Runner, error) {
 	runners, err := d.db.GetIdleRunnersForLanguage(ctx, language, int32(limit))
 	if isNotFound(err) {
 		return nil, nil
@@ -360,7 +488,7 @@ func (d *Postgres) GetIdleRunnersForLanguage(ctx context.Context, language strin
 }
 
 // GetRoutingTable returns the endpoints for all runners for the given module.
-func (d *Postgres) GetRoutingTable(ctx context.Context, module string) ([]string, error) {
+func (d *DAL) GetRoutingTable(ctx context.Context, module string) ([]string, error) {
 	routes, err := d.db.GetRoutingTable(ctx, module)
 	if len(routes) == 0 {
 		return nil, errors.WithStack(ErrNotFound)
@@ -368,7 +496,7 @@ func (d *Postgres) GetRoutingTable(ctx context.Context, module string) ([]string
 	return routes, errors.WithStack(translatePGError(err))
 }
 
-func (d *Postgres) GetRunnerState(ctx context.Context, runnerKey model.RunnerKey) (RunnerState, error) {
+func (d *DAL) GetRunnerState(ctx context.Context, runnerKey model.RunnerKey) (RunnerState, error) {
 	state, err := d.db.GetRunnerState(ctx, sqltypes.Key(runnerKey))
 	if err != nil {
 		return "", errors.WithStack(translatePGError(err))
@@ -376,12 +504,12 @@ func (d *Postgres) GetRunnerState(ctx context.Context, runnerKey model.RunnerKey
 	return RunnerState(state), nil
 }
 
-func (d *Postgres) ExpireRunnerClaims(ctx context.Context) (int64, error) {
+func (d *DAL) ExpireRunnerClaims(ctx context.Context) (int64, error) {
 	count, err := d.db.ExpireRunnerReservations(ctx)
 	return count, errors.WithStack(translatePGError(err))
 }
 
-func (d *Postgres) InsertDeploymentLogEntry(ctx context.Context, deployment model.DeploymentKey, logEntry log.Entry) error {
+func (d *DAL) InsertDeploymentLogEntry(ctx context.Context, deployment model.DeploymentKey, logEntry log.Entry) error {
 	logError := pgtype.Text{}
 	if logEntry.Error != nil {
 		logError.String = logEntry.Error.Error()
@@ -397,7 +525,7 @@ func (d *Postgres) InsertDeploymentLogEntry(ctx context.Context, deployment mode
 	})))
 }
 
-func (d *Postgres) loadDeployment(ctx context.Context, deployment sql.GetDeploymentRow) (*model.Deployment, error) {
+func (d *DAL) loadDeployment(ctx context.Context, deployment sql.GetDeploymentRow) (*model.Deployment, error) {
 	pm := &pschema.Module{}
 	err := proto.Unmarshal(deployment.Schema, pm)
 	if err != nil {
@@ -428,7 +556,7 @@ func (d *Postgres) loadDeployment(ctx context.Context, deployment sql.GetDeploym
 	return out, nil
 }
 
-func (d *Postgres) InsertMetricEntry(ctx context.Context, metric Metric) error {
+func (d *DAL) InsertMetricEntry(ctx context.Context, metric Metric) error {
 	var metricType sql.MetricType
 	switch metric.DataPoint.(type) {
 	case MetricHistogram:
@@ -458,7 +586,7 @@ func (d *Postgres) InsertMetricEntry(ctx context.Context, metric Metric) error {
 	})))
 }
 
-func (d *Postgres) GetMetricsForSourceModules(ctx context.Context, modules []string) ([]Metric, error) {
+func (d *DAL) GetMetricsForSourceModules(ctx context.Context, modules []string) ([]Metric, error) {
 	dbMetrics, err := d.db.GetMetricsBySourceModules(ctx, modules)
 	if err != nil {
 		return nil, errors.Wrap(translatePGError(err), "could not get metrics")
